--- src/components/aiServiceDetail/aiServiceDetailComment/aiServiceDetailTip/style.jsx
@@ -1,107 +1,0 @@
-import { styled } from "styled-components";
-import { BiSolidPencil } from "react-icons/bi";
-import { FaRegThumbsUp } from "react-icons/fa6";
-import { FaRegCommentAlt } from "react-icons/fa";
-
-export const AiServiceDetailTipWrap = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-`;
-
-// 헤더 (글쓰기, 정렬순)
-export const AiServiceDetailTipHeader = styled.div`
-  display: flex;
-  justify-content: space-around;
-  margin-top: 5.1rem;
-`;
-
-export const AiServiceDetailTipHeaderWrite = styled.div`
-  cursor: pointer;
-
-  :hover {
-    color: white;
-    background-color: #4285f4;
-  }
-`;
-
-export const AiServiceDetailTipHeaderWriteContent = styled.div`
-  display: flex;
-  font-size: 2rem;
-  font-weight: 500;
-  color: #4285f4;
-
-  border-radius: 10px;
-  border: 2px solid #4285f4;
-  padding: 0.7rem 1rem;
-  margin-right: 88rem;
-
-  :hover {
-    color: white;
-    background-color: #4285f4;
-  }
-`;
-
-export const StyledPencilIcon = styled(BiSolidPencil)`
-  font-size: 2rem;
-  margin-right: 0.5rem;
-`;
-
-export const AiServiceDetailTipHeaderSort = styled.div`
-  display: flex;
-`;
-
-export const AiServiceDetailTipLine = styled.div`
-  display: flex;
-  margin-top: 2.7rem;
-  background: #4285f4;
-  width: 110rem;
-  height: 2px;
-`;
-
-// 이용꿀팁 목록
-export const AiServiceDetailTipTable = styled.table`
-  margin: 0 auto;
-  width: 100%;
-`;
-
-export const AiServiceDetailTipTableThead = styled.thead``;
-export const AiServiceDetailTipTableTbody = styled.tbody``;
-
-export const AiServiceDetailTipTableTr = styled.tr``;
-export const AiServiceDetailTipTableTh = styled.th`
-  padding: 2rem;
-  font-size: 1.6rem;
-  border-bottom: 2px solid #f0f0f0;
-`;
-
-export const AiServiceDetailTipTableTd = styled.td`
-  padding: 2rem;
-  font-size: 1.6rem;
-  border-bottom: 2px solid #f0f0f0;
-  text-align: center;
-`;
-
-export const LikeIcon = styled(FaRegThumbsUp)`
-  color: black;
-  font-size: 1.4rem;
-  margin-right: 0.5rem;
-`;
-
-export const CommentIcon = styled(FaRegCommentAlt)`
-  color: black;
-  font-size: 1.4rem;
-  margin-right: 0.5rem;
-  margin-left: 0.5rem;
-`;
-
-export const AiServiceDetailTipPaging = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-top: 10rem;
-  position: relative;
-  bottom: 0;
-  transform: translate(-50%, 0);
-`;
